+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+UNRELEASED -
+-------------------
+
+## Added
+- Default to adding a `hart-role` label on GCE when creating minion from role.
+
+
 0.16.2 - 2023-11-01
 -------------------